requirements.txt
pyfiglet
requests
dnspython
python-whois
beautifulsoup4
colorama
rich
urllib3
holehe
